requirements.txt
fastapi
aiohttp
uvicorn
pydantic
azure.cosmos
scipy
python-dotenv
